refactor(platform): Phase 5a+5b — services layer, import hygiene, contextvars override
5a — MCP becomes a thin adapter over shared domain logic:
- New skill_seekers.services package: marketplace_manager,
  marketplace_publisher, config_publisher (sole home of detect_category /
  CATEGORY_KEYWORDS), source_manager, git_repo (moved too — it was the
  publishers' only mcp dependency, inverting it makes services
  mcp-extras-free, pinned by a regression test). CLI can now import this
  domain logic without depending on the optional mcp extras.
- Back-compat shims at the old mcp/ paths (named re-exports + identity
  tests); 29 in-repo import sites and mock patch targets updated.
- All 7 sys.path.insert import hacks in mcp/ removed (config_tools ×2,
  vector_db_tools, source_tools, server_fastmcp, server_legacy ×2) —
  replaced with absolute skill_seekers.* imports; server_fastmcp's
  100-line relative-import/sys.path fallback collapsed to one import
  block. No more dual module identities; packaged installs unaffected
  by cwd.

5b — ExecutionContext.override() is context-local:
- The override layer now lives in a ContextVar over the unchanged base
  singleton instead of mutating cls._instance: concurrent threads/asyncio
  tasks (the MCP server) can no longer clobber each other's overrides;
  nested overrides stack and unwind via tokens; exceptions restore;
  is_initialized() semantics unchanged. Thread propagation contract
  (copy_context) documented — the same pattern that already carries the
  MCP log-capture token. +6 concurrency/nesting/restore tests.

Remaining Phase 5 work (5c command migration + argument registry,
5d MCP context init + in-process dispatch) is scoped in
docs/UNIFICATION_PLAN.md.

Verified: full fast suite 3695 passed / 105 skipped (run post-5a),
combined 5a+5b suites green, ruff clean.
