+++ src/skill_seekers/mcp/tools/config_tools.py
@@ -6,18 +6,14 @@ for documentation scraping.
 """
 
 import json
-import sys
 from pathlib import Path
 
-from skill_seekers.mcp.tools._common import CLI_DIR, TextContent, text_response
+from skill_seekers.mcp.tools._common import TextContent, text_response
 
-
-# Path to CLI tools
 
 # Import config validator for validation
-sys.path.insert(0, str(CLI_DIR))
 try:
-    from config_validator import ConfigValidator
+    from skill_seekers.cli.config_validator import ConfigValidator
 except ImportError:
     ConfigValidator = None  # Graceful degradation if not available
 
@@ -178,9 +174,6 @@ async def validate_config(args: dict) -> list[TextContent]:
     """
     config_path = args["config_path"]
 
-    # Import validation classes
-    sys.path.insert(0, str(CLI_DIR))
-
     try:
         # Check if file exists
         if not Path(config_path).exists():
@@ -190,7 +183,7 @@ async def validate_config(args: dict) -> list[TextContent]:
 
         # Try unified config validator first
         try:
-            from config_validator import validate_config
+            from skill_seekers.cli.config_validator import validate_config
 
             validator = validate_config(config_path)
 
@@ -247,7 +240,7 @@ async def validate_config(args: dict) -> list[TextContent]:
             # Fall back to legacy validation
             import json
 
-            from doc_scraper import validate_config
+            from skill_seekers.cli.doc_scraper import validate_config
 
             with open(config_path) as f:
                 config = json.load(f)
