+++ tests/test_git_repo.py
@@ -10,7 +10,7 @@ from unittest.mock import MagicMock, patch
 import pytest
 from git.exc import GitCommandError
 
-from skill_seekers.mcp.git_repo import GitConfigRepo
+from skill_seekers.services.git_repo import GitConfigRepo
 
 
 @pytest.fixture
@@ -129,7 +129,7 @@ class TestInjectToken:
 class TestCloneOrPull:
     """Test clone and pull operations."""
 
-    @patch("skill_seekers.mcp.git_repo.git.Repo.clone_from")
+    @patch("skill_seekers.services.git_repo.git.Repo.clone_from")
     def test_clone_new_repo(self, mock_clone, git_repo):
         """Test cloning a new repository."""
         mock_clone.return_value = MagicMock()
@@ -147,7 +147,7 @@ class TestCloneOrPull:
         assert call_kwargs["single_branch"] is True
         assert call_kwargs["branch"] == "main"
 
-    @patch("skill_seekers.mcp.git_repo.git.Repo")
+    @patch("skill_seekers.services.git_repo.git.Repo")
     def test_pull_existing_repo(self, mock_repo_class, git_repo, temp_cache_dir):
         """Test pulling updates to existing repository."""
         # Create fake existing repo
@@ -168,7 +168,7 @@ class TestCloneOrPull:
         assert result == repo_path
         mock_origin.pull.assert_called_once_with("main")
 
-    @patch("skill_seekers.mcp.git_repo.git.Repo")
+    @patch("skill_seekers.services.git_repo.git.Repo")
     def test_pull_with_token_update(self, mock_repo_class, git_repo, temp_cache_dir):
         """Test pulling with token updates remote URL."""
         # Create fake existing repo
@@ -193,7 +193,7 @@ class TestCloneOrPull:
         updated_url = mock_origin.set_url.call_args[0][0]
         assert "ghp_token123@github.com" in updated_url
 
-    @patch("skill_seekers.mcp.git_repo.git.Repo.clone_from")
+    @patch("skill_seekers.services.git_repo.git.Repo.clone_from")
     def test_force_refresh_deletes_cache(self, mock_clone, git_repo, temp_cache_dir):
         """Test force refresh deletes existing cache."""
         # Create fake existing repo
@@ -211,7 +211,7 @@ class TestCloneOrPull:
         # Verify clone was called (not pull)
         mock_clone.assert_called_once()
 
-    @patch("skill_seekers.mcp.git_repo.git.Repo.clone_from")
+    @patch("skill_seekers.services.git_repo.git.Repo.clone_from")
     def test_clone_with_custom_branch(self, mock_clone, git_repo):
         """Test cloning with custom branch."""
         mock_clone.return_value = MagicMock()
@@ -228,7 +228,7 @@ class TestCloneOrPull:
         with pytest.raises(ValueError, match="Invalid git URL"):
             git_repo.clone_or_pull(source_name="test-source", git_url="not-a-valid-url")
 
-    @patch("skill_seekers.mcp.git_repo.git.Repo.clone_from")
+    @patch("skill_seekers.services.git_repo.git.Repo.clone_from")
     def test_clone_auth_failure_error(self, mock_clone, git_repo):
         """Test authentication failure error handling."""
         mock_clone.side_effect = GitCommandError(
@@ -240,7 +240,7 @@ class TestCloneOrPull:
                 source_name="test-source", git_url="https://github.com/org/repo.git"
             )
 
-    @patch("skill_seekers.mcp.git_repo.git.Repo.clone_from")
+    @patch("skill_seekers.services.git_repo.git.Repo.clone_from")
     def test_clone_not_found_error(self, mock_clone, git_repo):
         """Test repository not found error handling."""
         mock_clone.side_effect = GitCommandError("clone", 128, stderr="fatal: repository not found")
