+++ tests/test_mcp_git_sources.py
@@ -130,7 +130,7 @@ class TestFetchConfigModes:
             config_file = temp_dirs["dest"] / "react.json"
             assert config_file.exists()
 
-    @patch("skill_seekers.mcp.git_repo.GitConfigRepo")
+    @patch("skill_seekers.services.git_repo.GitConfigRepo")
     async def test_fetch_config_git_url_mode(self, mock_git_repo_class, temp_dirs):
         """Test Git URL mode - direct git clone."""
         from skill_seekers.mcp.server import fetch_config_tool
@@ -167,8 +167,8 @@ class TestFetchConfigModes:
         config_file = temp_dirs["dest"] / "react.json"
         assert config_file.exists()
 
-    @patch("skill_seekers.mcp.git_repo.GitConfigRepo")
-    @patch("skill_seekers.mcp.source_manager.SourceManager")
+    @patch("skill_seekers.services.git_repo.GitConfigRepo")
+    @patch("skill_seekers.services.source_manager.SourceManager")
     async def test_fetch_config_source_mode(
         self, mock_source_manager_class, mock_git_repo_class, temp_dirs
     ):
@@ -216,7 +216,7 @@ class TestFetchConfigModes:
         """Test error when source doesn't exist."""
         from skill_seekers.mcp.server import fetch_config_tool
 
-        with patch("skill_seekers.mcp.source_manager.SourceManager") as mock_sm_class:
+        with patch("skill_seekers.services.source_manager.SourceManager") as mock_sm_class:
             mock_sm = MagicMock()
             mock_sm.get_source.side_effect = KeyError("Source 'nonexistent' not found")
             mock_sm_class.return_value = mock_sm
@@ -228,7 +228,7 @@ class TestFetchConfigModes:
             assert "❌" in result[0].text
             assert "not found" in result[0].text
 
-    @patch("skill_seekers.mcp.git_repo.GitConfigRepo")
+    @patch("skill_seekers.services.git_repo.GitConfigRepo")
     async def test_fetch_config_config_not_found_in_repo(self, mock_git_repo_class, temp_dirs):
         """Test error when config doesn't exist in repository."""
         from skill_seekers.mcp.server import fetch_config_tool
@@ -252,7 +252,7 @@ class TestFetchConfigModes:
         assert "not found" in result[0].text
         assert "Available configs" in result[0].text
 
-    @patch("skill_seekers.mcp.git_repo.GitConfigRepo")
+    @patch("skill_seekers.services.git_repo.GitConfigRepo")
     async def test_fetch_config_invalid_git_url(self, mock_git_repo_class):
         """Test error handling for invalid git URL."""
         from skill_seekers.mcp.server import fetch_config_tool
@@ -279,7 +279,7 @@ class TestSourceManagementTools:
         """Test adding a new config source."""
         from skill_seekers.mcp.server import add_config_source_tool
 
-        with patch("skill_seekers.mcp.source_manager.SourceManager") as mock_sm_class:
+        with patch("skill_seekers.services.source_manager.SourceManager") as mock_sm_class:
             mock_sm = MagicMock()
             mock_sm.add_source.return_value = {
                 "name": "team",
@@ -332,7 +332,7 @@ class TestSourceManagementTools:
         """Test error when source name is invalid."""
         from skill_seekers.mcp.server import add_config_source_tool
 
-        with patch("skill_seekers.mcp.source_manager.SourceManager") as mock_sm_class:
+        with patch("skill_seekers.services.source_manager.SourceManager") as mock_sm_class:
             mock_sm = MagicMock()
             mock_sm.add_source.side_effect = ValueError(
                 "Invalid source name 'team@company'. Must be alphanumeric with optional hyphens/underscores."
@@ -350,7 +350,7 @@ class TestSourceManagementTools:
         """Test listing config sources."""
         from skill_seekers.mcp.server import list_config_sources_tool
 
-        with patch("skill_seekers.mcp.source_manager.SourceManager") as mock_sm_class:
+        with patch("skill_seekers.services.source_manager.SourceManager") as mock_sm_class:
             mock_sm = MagicMock()
             mock_sm.list_sources.return_value = [
                 {
@@ -389,7 +389,7 @@ class TestSourceManagementTools:
         """Test listing when no sources registered."""
         from skill_seekers.mcp.server import list_config_sources_tool
 
-        with patch("skill_seekers.mcp.source_manager.SourceManager") as mock_sm_class:
+        with patch("skill_seekers.services.source_manager.SourceManager") as mock_sm_class:
             mock_sm = MagicMock()
             mock_sm.list_sources.return_value = []
             mock_sm_class.return_value = mock_sm
@@ -404,7 +404,7 @@ class TestSourceManagementTools:
         """Test listing only enabled sources."""
         from skill_seekers.mcp.server import list_config_sources_tool
 
-        with patch("skill_seekers.mcp.source_manager.SourceManager") as mock_sm_class:
+        with patch("skill_seekers.services.source_manager.SourceManager") as mock_sm_class:
             mock_sm = MagicMock()
             mock_sm.list_sources.return_value = [
                 {
@@ -433,7 +433,7 @@ class TestSourceManagementTools:
         """Test removing a config source."""
         from skill_seekers.mcp.server import remove_config_source_tool
 
-        with patch("skill_seekers.mcp.source_manager.SourceManager") as mock_sm_class:
+        with patch("skill_seekers.services.source_manager.SourceManager") as mock_sm_class:
             mock_sm = MagicMock()
             mock_sm.remove_source.return_value = True
             mock_sm_class.return_value = mock_sm
@@ -453,7 +453,7 @@ class TestSourceManagementTools:
         """Test removing non-existent source."""
         from skill_seekers.mcp.server import remove_config_source_tool
 
-        with patch("skill_seekers.mcp.source_manager.SourceManager") as mock_sm_class:
+        with patch("skill_seekers.services.source_manager.SourceManager") as mock_sm_class:
             mock_sm = MagicMock()
             mock_sm.remove_source.return_value = False
             mock_sm.list_sources.return_value = [
@@ -488,8 +488,8 @@ class TestSourceManagementTools:
 class TestCompleteWorkflow:
     """Test complete workflow of add → fetch → remove."""
 
-    @patch("skill_seekers.mcp.git_repo.GitConfigRepo")
-    @patch("skill_seekers.mcp.source_manager.SourceManager")
+    @patch("skill_seekers.services.git_repo.GitConfigRepo")
+    @patch("skill_seekers.services.source_manager.SourceManager")
     async def test_add_fetch_remove_workflow(self, mock_sm_class, mock_git_repo_class, temp_dirs):
         """Test complete workflow: add source → fetch config → remove source."""
         from skill_seekers.mcp.server import (
